@@ -7,14 +7,14 @@
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@types/jest": "24.0.23",
     "@types/node": "12.12.14",
-    "@types/react": "16.9.14",
-    "@types/react-dom": "16.9.4",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
     "@types/react-router-dom": "^5.3.3",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-hook-form": "^7.33.1",
     "react-router-dom": "^6.3.0",
-    "react-scripts": "3.2.0",
+    "react-scripts": "^5.0.1",
     "react-select": "^5.4.0",
     "sass": "^1.54.0",
     "sort-by-typescript": "^1.2.3",
